+++ run-nextflow-tests/entry.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"Quick entrypoint for this tool."
+"Dockerfile entrypoint script."
 import argparse
 import sys
 from pathlib import Path
@@ -7,20 +7,25 @@ from pathlib import Path
 from configtest import NextflowConfigTest
 
 
-def run_pipeline_test(pipeline: Path, test_case: Path):
-    "Run the bundled pipeline-recalibrate-BAM test."
-    testobj = NextflowConfigTest.from_file(test_case)
+def run_pipeline_test(pipeline: Path, test_case: Path) -> bool:
+    "Run a pipeline test and save the results to a new file."
+    testobj = NextflowConfigTest.from_file(pipeline, test_case)
+    updated_testobj = testobj.recompute_results()
 
-    if testobj.check_results(pipeline):
-        print("No changes!")
-    else:
-        sys.exit(1)
+    # Print any differences
+    testobj.print_diffs(updated_testobj)
+    updated_testobj.mark_for_archive()
+
+    return testobj == updated_testobj
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument("pipeline_path")
-    parser.add_argument("test_path")
+    parser.add_argument("pipeline_path", type=Path)
+    parser.add_argument("test_path", type=Path)
 
     args = parser.parse_args()
-    run_pipeline_test(Path(args.pipeline_path).resolve(), Path(args.test_path))
+    if run_pipeline_test(args.pipeline_path.resolve(), args.test_path):
+        sys.exit(0)
+
+    sys.exit(1)